fix: pr comments, move tests
